requirements.txt
apache-airflow-providers-amazon==8.16.0
apache-airflow-providers-postgres==5.10.0
python-dotenv==1.0.0
pyyaml==6.0.1
boto3==1.34.0
psycopg2-binary==2.9.9
minio==7.2.0
pypdf2==3.0.1
pdf2image==1.16.3
pillow==10.2.0
pydantic==2.8.2
openai==1.35.0
anthropic==0.31.0
google-generativeai==0.7.0
langfuse==3.3.0
tenacity==8.3.0
streamlit==1.35.0
